a C program that prints the size of various types on the computer it is compiled and run on

+++ 0x00-hello_world/6-size.c
@@ -1,4 +1,4 @@
-#include<stdio.h>
+#include <stdio.h>
 /**
 *main - Entry point
 *return: (0)
@@ -11,11 +11,12 @@ int main(void)
 	long long int d;
 	float f;
 
-	printf("Size of a char: %lu byte(s)\n" (unsigned long)sizeof(a));
-	printf("Size of an int: %lu byte(s)\n" (unsigned long)sizeof(b));
-	printf("Size of a long int: %lu byte(s)\n" (unsigned long)sizeof(c));
-	printf("Size of a long long int: %lu byte(s)\n" (unsigned long)sizeof(d));
-	printf("Size of a float: %lu bytes(s)\n" (unsigned long)sizeof(f));
+	//sizeof evaluates the size of the data type
+	printf("Size of a char: %lu byte(s)\n", sizeof(a));
+	printf("Size of an int: %lu byte(s)\n", sizeof(b));
+	printf("Size of a long int: %lu byte(s)\n", sizeof(c));
+	printf("Size of a long long int: %lu byte(s)\n", sizeof(d));
+	printf("Size of a float: %lu bytes(s)\n", sizeof(f));
 	return (0);
 }
